Стру-ция состояния. Изменения ссылок на со-ие.
