+++ src/components/12 Profile/ProfileNav.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { navAC } from "../../store/actions";
+
 
 
 //Компанента отвечает за отображение интерфейса входа
