+++ src/components/31 SettingsMaster/311 MasterSpecialization/SpecializationContainer.js
@@ -4,14 +4,14 @@ import { connect } from "react-redux";
 import SpecializationAPIContainer from "./SpecializationAPIContainer";
 
 //Экшены
-import { transferDataSpecializationAC } from "../../../store/actions";
+import { transferDataSpecializationAC } from "../../../store/3 Master/31actions";
 
 
 
 let mapStateToProps = (state) => {
     return{
-        SID: state.reduserMain.SID,                                         //Сессия
-        specializations: state.rediserSettings.specializations,             //Специализации
+        SID: state.rGeneral.SID,                                         //Сессия
+        specializations: state.rMaster.specializations,             //Специализации
         
     }
 }
