+++ src/components/32 ScheduleMaster/ScheduleContainer.js
@@ -4,16 +4,13 @@ import { connect } from "react-redux";
 import ScheduleAPIContainer from "./ScheduleAPIContainer";
 
 //Экшены
-import { navAC } from "../../store/actions";
-import { transferInfoDateAC } from "../../store/actions";
+import { transferInfoDateAC } from "../../store/3 Master/32actions";
 
 
 
 let mapStateToProps = (state) => {
     return{
-        //textEror: state.reduserMain.textEror,   //Текст ошибки
-        nav: state.reduserMain.nav,                 //Навигация
-        user: state.reduserMain.user,               //Данные пользователя
+        user: state.rGeneral.user,               //Данные пользователя
     }
 }
 
@@ -21,18 +18,10 @@ let mapStateToProps = (state) => {
 let mapDispatchToProps = (dispatch ) => {
 
     return{
-        //Навигация
-        navAC: (path) =>{
-            dispatch(navAC(path))
-        }, 
-
         //Передает данные о дате
         transferInfoDateAC: (date) =>{
             dispatch(transferInfoDateAC(date))
         },
-
-        
-        
     }
 }
