+++ src/components/41 Reception/412 ChoiseMaster/4121 DatePlanClient/DatePlanContainer.js
@@ -4,16 +4,16 @@ import { connect } from "react-redux";
 import DatePlanAPIContainer from "./DatePlanAPIContainer";
 
 //Экшены
-import { trDataDateAC } from "../../../../store/Client/actions";
+import { trDataDateAC } from "../../../../store/4 Client/41actions";
 
 
 let mapStateToProps = (state) => {
     return{
-        SID: state.reduserMain.SID,                         //Сессия
+        SID: state.rGeneral.SID,                         //Сессия
         
-        selectDate: state.rediserSchedule.selectDate,       //Выбранная дата
-        idSelectMaster: state.clientR.idSelectServ,         //Хранит id выбранного мастера
-        arrDataDate: state.clientR.arrDataDate,             //Хранит информацию для плана даты
+        selectDate: state.rClient.selectDate,       //Выбранная дата
+        idSelectMaster: state.rClient.idSelectMaster,         //Хранит id выбранного мастера
+        arrDataDate: state.rClient.arrDataDate,             //Хранит информацию для плана даты
         
         
     }
