+++ src/store/1 General/actionType.js
@@ -1,0 +1,19 @@
+//===== App =====
+export const GET_INFO_PROFILE = 'GET_INFO_PROFILE'                  //Передает информацию о пользователе
+
+
+//===== 1 =====
+export const NAV = 'NAV'                        //Навигация
+export const TEXT_EROR = 'TEXT_EROR'            //Текст ошибки
+export const GET_SID = 'GET_SID'                //Передает сессию
+
+
+
+//=====   =====
+export const DATA_POP_UP = 'DATA_POP_UP'        //Передает данные о всплывающем окне
+
+
+
+//===== 14 =====
+export const GET_USERS = 'GET_USERS'                            //Передает список пользователей
+export const GET_STATUSES = 'GET_STATUSES'
