+++ src/store/2 Boss/actionType.js
@@ -1,0 +1,11 @@
+
+
+
+//============ 211 ================
+export const TRANSFER_DATA_SPECIALIZATION = 'TRANSFER_DATA_SPECIALIZATION'      //Передает данные о имеющихся специализациях
+
+
+
+//===== 212 =====
+export const TRANS_DATA_SERVICES = 'TRANS_DATA_SERVICES'                        //Передает данные о имеющихся услугах
+export const TRANS_DATA_SPEC_SERVICES = 'TRANS_DATA_SPEC_SERVICES'              //Передает данные для объедененного массива специализаций и услуг
