+++ src/store/3 Master/31actions.js
@@ -1,0 +1,16 @@
+import * as actions from './actionType';
+
+//============ 311 ================
+export const transferDataSpecializationAC = (arrData) => ({
+    type: actions.TRANSFER_DATA_SPECIALIZATION,
+    payload: arrData
+})
+
+
+
+//========== 312 ==========
+//Данные для настройки услуг
+export const trDataForSetingServiceAC = (arrData) => ({
+    type: actions.TRANSP_DATA_FOR_SETING_SERVICE,
+    payload: arrData,
+})
