+++ src/store/3 Master/32actions.js
@@ -1,0 +1,8 @@
+import * as actions from './actionType';
+
+//========= 32 ============
+//Передает данные о выбранной дате
+export const transferInfoDateAC = (date) => ({
+    type: actions.TRANSFER_INFO_DATE,
+    payload: date
+})
