+++ src/store/3 Master/actionType.js
@@ -1,0 +1,23 @@
+
+
+
+//============ 311 ================
+export const TRANSFER_DATA_SPECIALIZATION = 'TRANSFER_DATA_SPECIALIZATION'      //Передает данные о имеющихся специализациях
+
+
+//========== 312 ==========
+export const TRANSP_DATA_FOR_SETING_SERVICE = 'TRANSP_DATA_FOR_SETING_SERVICE'      //Данные для настройки услуг
+
+
+//========== 33 ===========
+export const TRANSFER_INFO_DATE = 'TRANSFER_INFO_DATE'          //Передает данные о выбранной дате
+export const TRANSP_DATA_DATE = 'TRANSP_DATA_DATE'          //Передает данные из базы расписаний на выбранную дату
+export const TRANSP_DATA_FORM_DATE_PLAN = 'TRANSP_DATA_FORM_DATE_PLAN'              //Данные для формы окна даты
+
+
+
+
+
+
+
+
