+++ src/store/4 Client/actionType.js
@@ -1,13 +1,23 @@
 
+//============ 41 ================
 export const CLEAN_STATE = 'CLEAN_STATE'            //Очищает состояние
 
-//ВЫБОР УСЛУГИ
+
+//========== 411 ==========
 export const TR_SPECIALIZATION_WITH_SERVICE = 'TR_SPECIALIZATION_WITH_SERVICE'      //Данные для окна выбора услуг
 export const TR_FREE_DATES = 'TR_FREE_DATES'                                        //Свободные даты на выбранную услугу
 export const TR_ID_SERVECE = 'TR_ID_SERVECE'                                        //Передает id выбранной усдуги
-//ВЫБОР МАСТЕРА
+
+
+//========== 4111 ==========
+export const TRANSFER_INFO_DATE = 'TRANSFER_INFO_DATE'          //Передает данные о выбранной дате
+export const TR_DATA_DATE = 'TR_DATA_DATE'          //Передает информацию для плана даты
+
+
+//========== 412 ==========
 export const TR_DATA_MASTERS = 'TR_DATA_MASTERS'            //Передает информацию о мастерах
 export const TR_DATA_FOR_CALENDAR = 'TR_DATA_FOR_CALENDAR'  //Передает информацию для календаря
 
-//ДАТА ПЛАН
-export const TR_DATA_DATE = 'TR_DATA_DATE'          //Передает информацию для плана даты
+
+//========== 4121 ==========
+export const TR_ID_MASTER = 'TR_ID_MASTER'                                        //Передает id выбранной усдуги
