--- src/store/Master/actionType.js
@@ -1,9 +1,0 @@
-
-//============= ДЛЯ НАСТРОЙКИ УСЛУГ ==============
-export const TRANSP_DATA_FOR_SETING_SERVICE = 'TRANSP_DATA_FOR_SETING_SERVICE'      //Данные для настройки услуг
-export const TRANSP_DATA_FORM_DATE_PLAN = 'TRANSP_DATA_FORM_DATE_PLAN'              //Данные для формы окна даты
-
-//================ ДЛЯ ОКНА ДАТЫ =================
-export const TRANSP_DATA_DATE = 'TRANSP_DATA_DATE'          //Передает данные из базы расписаний на выбранную дату
-
-
