--- src/store/actionType.js
@@ -1,39 +1,0 @@
-//======================= reduserMain =================
-
-export const NAV = 'NAV'                        //Навигация
-export const TEXT_EROR = 'TEXT_EROR'            //Текст ошибки
-export const GET_SID = 'GET_SID'                //Передает сессию
-export const GET_INFO_PROFILE = 'GET_INFO_PROFILE'                  //Передает информацию о пользователе
-export const DATA_POP_UP = 'DATA_POP_UP'                            //Передает данные о всплывающем окне
-
-
-//======================= reduserMain end =============
-
-//======================= USERS =================
-export const GET_USERS = 'GET_USERS'                            //Передает список пользователей
-export const GET_STATUSES = 'GET_STATUSES'                      //Передает список статусов пользователей
-//======================= USERS end =============
-
-
-//======================= Schedule ==========================
-export const TRANSFER_INFO_DATE = 'TRANSFER_INFO_DATE'          //Передает данные о выбранной дате
-export const TRANSFER_DATA_DATE = 'TRANSFER_DATA_DATE'          //Передает данные из базы расписаний на выбранную дату
-//======================= Schedule end ======================
-
-
-//======================= Settings ==========================
-export const TRANSFER_DATA_SPECIALIZATION = 'TRANSFER_DATA_SPECIALIZATION'      //Передает данные о имеющихся специализациях
-export const TRANS_DATA_SERVICES = 'TRANS_DATA_SERVICES'                        //Передает данные о имеющихся услугах
-export const TRANS_DATA_SPEC_SERVICES = 'TRANS_DATA_SPEC_SERVICES'              //Передает данные для объедененного массива специализаций и услуг
-export const TR_DATA_SERVICES_FOR_SPEC = 'TR_DATA_SERVICES_FOR_SPEC'            //Передает данные для массива для формы специалтиз
-
-//======================= Settings end ==========================
-
-//export const GET_SID = 'GET_SID'                        //Получение сессии
-
-
-
-
-
-
-
